Cleanup `simulators` and `trajectories`
Simulators and trajectories have been cleaned up.
The old folders have been deleted and the suffix 'hierarchical' has been
removed from the current implementation.
A `zero_simulator.py` module has been added,
which implements a simulator with zero dynamics.
The conversion functions have been moved outside the `Simulator` class.
(We can decide if we want them outside or inside the `Simulator` class,
but in the last commit there was one inside and one outside.)
The changes were tested on the GUI.
There is some trouble with the module `attitude_inner_loop_simulator.py`,
but from a quick look I think it is still in the making.
I added a few lines to it, but I think it needs more fixes.