Parameter serialization and basic plots.
The class `Jsonable` now has a method to return a string representation
of the parameters (for example, the gains). The method uses the usual
`json.dumps` function.

The `Mission` class now has a `plot_from_string` function,
that draws 4 plots based on a string that contains serialized numerical data.
The child missions need to redefine this function
so that the number and content of the plots agrees with the
particular mission.